Fixed: build warnings (#2d932bv)
