+++ src/store/modules/order/actions.ts
@@ -108,121 +108,6 @@ const actions: ActionTree<OrderState , RootState> ={
     return resp;
   },
 
-  async deliverShipment ({ dispatch }, order) {
-    emitter.emit("presentLoader");
-
-    const params = {
-      shipmentId: order.shipmentId,
-      statusId: 'SHIPMENT_SHIPPED'
-    }
-
-    let resp;
-
-    try {
-      resp = await OrderService.updateShipment(params)
-      if (resp.status === 200 && !hasError(resp)) {
-        showToast(translate('Order delivered to', {customerName: order.customerName}))
-      } else {
-        showToast(translate("Something went wrong"))
-      }
-      emitter.emit("dismissLoader")
-    } catch(err) {
-      console.error(err)
-      showToast(translate("Something went wrong"))
-    }
-
-    emitter.emit("dismissLoader")
-    return resp;
-  },
-
-  async packDeliveryItems ({ commit }, shipmentId) {
-    const params = {
-      shipmentId: shipmentId,
-      statusId: 'SHIPMENT_PACKED'
-    }
-    return await OrderService.updateShipment(params)
-  },
-
-  async quickShipEntireShipGroup ({ dispatch }, payload) {
-    emitter.emit("presentLoader")
-
-    const params = {
-      orderId: payload.order.orderId,
-      setPackedOnly: 'Y',
-      dimensionUomId: 'WT_kg',
-      shipmentBoxTypeId: 'YOURPACKNG',
-      weight: '1',
-      weightUomId: 'WT_kg',
-      facilityId: payload.facilityId,
-      shipGroupSeqId: payload.shipGroupSeqId
-    }
-    
-    let resp;
-
-    try {
-      resp = await OrderService.quickShipEntireShipGroup(params)
-      if (resp.status === 200 && !hasError(resp) && resp.data._EVENT_MESSAGE_) {
-        /* To display the button label as per the shipmentMethodTypeId, this will only used on orders segment.
-          Because we get the shipmentMethodTypeId on items level in wms-orders API.
-          As we already get shipmentMethodTypeId on order level in readytoshiporders API hence we will not use this method on packed orders segment.
-        */
-        const shipmentMethodTypeId = payload.order.items.find((ele: any) => ele.shipGroupSeqId == payload.shipGroupSeqId).shipmentMethodTypeId
-        if (shipmentMethodTypeId !== 'STOREPICKUP') {
-          // TODO: find a better way to get the shipmentId
-          const shipmentId = resp.data._EVENT_MESSAGE_.match(/\d+/g)[0]
-          await dispatch('packDeliveryItems', shipmentId).then((data) => {
-            if (!hasError(data) && !data.data._EVENT_MESSAGE_) showToast(translate("Something went wrong"))
-          })
-        }
-        showToast(translate("Order packed and ready for delivery"))
-      } else {
-        showToast(translate("Something went wrong"))
-      }
-      emitter.emit("dismissLoader")
-    } catch(err) {
-      console.error(err)
-      showToast(translate("Something went wrong"))
-    }
-
-    emitter.emit("dismissLoader")
-    return resp;
-  },
-
-  // TODO: handle the unfillable items count
-  async setUnfillableOrderOrItem ({ dispatch }, payload) {
-    emitter.emit("presentLoader");
-    return await dispatch("rejectOrderItems", payload).then((resp) => {
-      const refreshPickupOrders = resp.find((response: any) => !(response.data._ERROR_MESSAGE_ || response.data._ERROR_MESSAGE_LIST_))
-      if (refreshPickupOrders) {
-        showToast(translate('All items were canceled from the order') + ' ' + payload.orderId);
-      } else {
-        showToast(translate('Something went wrong'));
-      }
-      emitter.emit("dismissLoader");
-      return resp;
-    }).catch(err => err);
-  },
-
-  rejectOrderItems ({ commit }, data) {
-    const payload = {
-      'orderId': data.orderId
-    }
-
-    return Promise.all(data.items.map((item: any) => {
-      const params = {
-        ...payload,
-        'rejectReason': item.reason,
-        'facilityId': item.facilityId,
-        'orderItemSeqId': item.orderItemSeqId,
-        'shipmentMethodTypeId': item.shipmentMethodTypeId,
-        'quantity': parseInt(item.quantity)
-      }
-      return OrderService.rejectOrderItem({'payload': params}).catch((err) => { 
-        return err;
-      })
-    }))
-  },
-
   // clearning the orders state when logout, or user store is changed
   clearOrders ({ commit }) {
     commit(types.ORDER_OPEN_UPDATED, {orders: {} , total: 0})
