+++ src/store/modules/product/actions.ts
@@ -3,9 +3,7 @@ import { ActionTree } from 'vuex'
 import RootState from '@/store/RootState'
 import ProductState from './ProductState'
 import * as types from './mutation-types'
-import { hasError, showToast } from '@/utils'
-import { translate } from '@/i18n'
-import emitter from '@/event-bus'
+import { hasError } from '@/utils'
 
 
 const actions: ActionTree<ProductState, RootState> = {
